Move eviction tests to a class of their own
