+++ methanol/src/test/java/com/github/mizosoft/methanol/internal/cache/StoreTest.java
@@ -33,7 +33,6 @@ import static com.github.mizosoft.methanol.internal.cache.StoreTesting.writeData
 import static com.github.mizosoft.methanol.internal.cache.StoreTesting.writeEntry;
 import static com.github.mizosoft.methanol.testing.TestUtils.awaitUninterruptibly;
 import static com.github.mizosoft.methanol.testing.junit.ExecutorExtension.ExecutorType.CACHED_POOL;
-import static com.github.mizosoft.methanol.testing.junit.StoreSpec.Execution.SAME_THREAD;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
@@ -48,7 +47,6 @@ import com.github.mizosoft.methanol.testing.junit.ExecutorExtension.ExecutorConf
 import com.github.mizosoft.methanol.testing.junit.StoreContext;
 import com.github.mizosoft.methanol.testing.junit.StoreExtension;
 import com.github.mizosoft.methanol.testing.junit.StoreExtension.StoreParameterizedTest;
-import com.github.mizosoft.methanol.testing.junit.StoreSpec;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
@@ -59,6 +57,7 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.CyclicBarrier;
 import java.util.concurrent.Executor;
 import java.util.concurrent.atomic.AtomicBoolean;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -225,6 +224,7 @@ class StoreTest {
     assertThat(store.size()).isEqualTo(sizeOf("Mew", "Mewtwo"));
   }
 
+  @Disabled("Till spec change")
   @StoreParameterizedTest
   void discardEditAfterRemove(Store store, StoreContext context) throws IOException {
     try (var editor = edit(store, "e1")) {
@@ -308,6 +308,7 @@ class StoreTest {
   }
 
   /** Removing an entry discards any ongoing edit for this entry. */
+  @Disabled("Till spec change")
   @StoreParameterizedTest
   void removeBeforeCommittingFirstEdit(Store store, StoreContext context) throws IOException {
     try (var editor = edit(store, "e1")) {
@@ -321,6 +322,7 @@ class StoreTest {
     assertThat(store.size()).isZero();
   }
 
+  @Disabled("Till spec change")
   @StoreParameterizedTest
   void clearWhileEditing(Store store, StoreContext context) throws IOException {
     writeEntry(store, "e1", "Jynx", "Raichu");
@@ -437,6 +439,7 @@ class StoreTest {
     }
   }
 
+  @Disabled("Till spec change")
   @StoreParameterizedTest
   void removeWhileEditingFromViewer(Store store, StoreContext context) throws IOException {
     writeEntry(store, "e1", "Pickachu", "Mewtwo");
@@ -466,6 +469,7 @@ class StoreTest {
     assertAbsent(store, context, "e1");
   }
 
+  @Disabled("Till spec change")
   @StoreParameterizedTest
   void removeFromViewerWhileEditingFromViewer(Store store, StoreContext context)
       throws IOException {
@@ -616,185 +620,6 @@ class StoreTest {
     }
   }
 
-  @StoreParameterizedTest
-  @StoreSpec(maxSize = 10, execution = SAME_THREAD)
-  void writeExactlyMaxSizeBytesByOneEntry(Store store) throws IOException {
-    writeEntry(store, "e1", "12345", "abcde"); // Grow size to 10 bytes.
-    assertEntryEquals(store, "e1", "12345", "abcde");
-    assertThat(store.size()).isEqualTo(10);
-  }
-
-  @StoreParameterizedTest
-  @StoreSpec(maxSize = 10, execution = SAME_THREAD)
-  void writeExactlyMaxSizeBytesByTwoEntries(Store store) throws IOException {
-    writeEntry(store, "e1", "12", "abc"); // Grow size to 5 bytes.
-    writeEntry(store, "e2", "45", "def"); // Grow size to 10 bytes.
-    assertEntryEquals(store, "e1", "12", "abc");
-    assertEntryEquals(store, "e2", "45", "def");
-    assertThat(store.size()).isEqualTo(10);
-  }
-
-  @StoreParameterizedTest
-  @StoreSpec(maxSize = 15, execution = SAME_THREAD)
-  void writeBeyondMaxSize(Store store, StoreContext context) throws IOException {
-    writeEntry(store, "e1", "12", "abc"); // Grow size to 5 bytes.
-    writeEntry(store, "e2", "34", "def"); // Grow size to 10 bytes.
-    assertThat(store.size()).isEqualTo(10);
-
-    // LRU queue: e2, e1.
-    view(store, "e1").close();
-
-    // Grow size to 16 bytes, causing e2 to be evicted.
-    writeEntry(store, "e3", "567", "ghi");
-
-    // LRU queue: e1, e3.
-    assertAbsent(store, context, "e2");
-    assertEntryEquals(store, "e1", "12", "abc");
-    assertEntryEquals(store, "e3", "567", "ghi");
-    assertThat(store.size()).isEqualTo(11);
-
-    // Grows size to 11 + 14 bytes causing both e1 & e3 to be evicted.
-    writeEntry(store, "e4", "Jynx", "Charmander");
-    assertAbsent(store, context, "e1");
-    assertAbsent(store, context, "e3");
-    assertEntryEquals(store, "e4", "Jynx", "Charmander");
-    assertThat(store.size()).isEqualTo(14);
-  }
-
-  @StoreParameterizedTest
-  @StoreSpec(maxSize = 15, execution = SAME_THREAD)
-  void discardedWriteBeyondMaxSize(Store store, StoreContext context) throws IOException {
-    writeEntry(store, "e1", "123", "abc"); // Grow size to 6 bytes.
-    writeEntry(store, "e2", "456", "def"); // Grow size to 12 bytes.
-    assertThat(store.size()).isEqualTo(12);
-
-    try (var editor = edit(store, "e3")) {
-      writeEntry(editor, "alpha", "beta");
-    }
-    assertAbsent(store, context, "e3");
-    assertEntryEquals(store, "e1", "123", "abc");
-    assertEntryEquals(store, "e2", "456", "def");
-    assertThat(store.size()).isEqualTo(12);
-  }
-
-  @StoreParameterizedTest
-  @StoreSpec(maxSize = 15, execution = SAME_THREAD)
-  void discardedByRemovalWriteBeyondMaxSize(Store store, StoreContext context) throws IOException {
-    writeEntry(store, "e1", "123", "abc"); // Grow size to 6 bytes.
-    writeEntry(store, "e2", "456", "def"); // Grow size to 12 bytes.
-    assertThat(store.size()).isEqualTo(12);
-
-    try (var editor = edit(store, "e3")) {
-      writeEntry(editor, "alpha", "beta");
-      editor.commitOnClose();
-
-      assertThat(store.remove("e3")).isTrue();
-    }
-    assertAbsent(store, context, "e3");
-    assertEntryEquals(store, "e1", "123", "abc");
-    assertEntryEquals(store, "e2", "456", "def");
-    assertThat(store.size()).isEqualTo(12);
-  }
-
-  @StoreParameterizedTest
-  @StoreSpec(maxSize = 14, execution = SAME_THREAD)
-  void writeBeyondMaxSizeByMetadataUpdate(Store store, StoreContext context) throws IOException {
-    writeEntry(store, "e1", "123", "abc"); // Grow size to 6 bytes.
-    writeEntry(store, "e2", "456", "def"); // Grow size to 12 bytes.
-    assertThat(store.size()).isEqualTo(12);
-
-    try (var editor = edit(store, "e1")) {
-      // Increase metadata by 3 bytes, causing size to grow to 15 bytes & e2 to be evicted.
-      setMetadata(editor, "123456");
-      editor.commitOnClose();
-    }
-    assertAbsent(store, context, "e2");
-    assertEntryEquals(store, "e1", "123456", "abc");
-    assertThat(store.size()).isEqualTo(9);
-  }
-
-  @StoreParameterizedTest
-  @StoreSpec(maxSize = 14, execution = SAME_THREAD)
-  void writeBeyondMaxSizeByDataUpdate(Store store, StoreContext context) throws IOException {
-    writeEntry(store, "e1", "123", "abc"); // Grow size to 6 bytes.
-    writeEntry(store, "e2", "456", "def"); // Grow size to 12 bytes.
-    assertThat(store.size()).isEqualTo(12);
-
-    try (var editor = edit(store, "e1")) {
-      // Increase data by 3 bytes, causing size to grow to 15 bytes & e2 to be evicted.
-      writeData(editor, "abcdef");
-      editor.commitOnClose();
-    }
-    assertAbsent(store, context, "e2");
-    assertEntryEquals(store, "e1", "123", "abcdef");
-    assertThat(store.size()).isEqualTo(9);
-  }
-
-  @StoreParameterizedTest
-  @StoreSpec(maxSize = 18, execution = SAME_THREAD)
-  void lruEviction(Store store, StoreContext context) throws IOException {
-    // Grow size to 6 bytes.
-    // LRU queue: e1.
-    writeEntry(store, "e1", "aaa", "bbb");
-    assertThat(store.size()).isEqualTo(6);
-
-    // Grow size to 12 bytes.
-    // LRU queue: e1, e2.
-    writeEntry(store, "e2", "ccc", "ddd");
-    assertThat(store.size()).isEqualTo(12);
-
-    // LRU queue: e2, e1.
-    view(store, "e1").close();
-
-    // Grow size to 18 bytes.
-    // LRU queue: e2, e1, e3.
-    writeEntry(store, "e3", "eee", "fff");
-    assertThat(store.size()).isEqualTo(18);
-
-    // LRU queue: e2, e3, e1.
-    view(store, "e1").close();
-
-    // Grow size to 24 bytes, causing e2 to be evicted to get down to 18.
-    // LRU queue: e3, e1, e4.
-    writeEntry(store, "e4", "ggg", "hhh");
-    assertAbsent(store, context, "e2");
-    assertThat(store.size()).isEqualTo(18);
-
-    // LRU queue: e1, e4, e3.
-    view(store, "e3").close();
-
-    // Grow size to 24 bytes, causing e1 to be evicted to get down to 18 bytes.
-    // LRU queue: e4, e3, e5
-    writeEntry(store, "e5", "iii", "jjj");
-    assertAbsent(store, context, "e1");
-    assertThat(store.size()).isEqualTo(18);
-
-    // Grow size to 18 + 12 bytes, causing e4 & e3 to be evicted to get down to 18 bytes.
-    // LRU queue: e5, e6.
-    writeEntry(store, "e6", "kkk", "lmnopqrst");
-    assertAbsent(store, context, "e4", "e3");
-    assertThat(store.size()).isEqualTo(18);
-
-    // LRU queue: e6, e5.
-    view(store, "e5").close();
-
-    // Grow size to 24 bytes, causing e6 to be evicted to get down to 12.
-    // LRU queue: e5, e7.
-    writeEntry(store, "e7", "uuu", "vvv");
-    assertAbsent(store, context, "e6");
-    assertThat(store.size()).isEqualTo(12);
-
-    // Grow size to 18 bytes, causing nothing to be evicted since size is within bounds.
-    // LRU queue: e5, e7, e8.
-    writeEntry(store, "e8", "xxx", "~!@");
-    assertThat(store.size()).isEqualTo(18);
-
-    // Write one 18 bytes entry, causing all other entries to be evicted.
-    writeEntry(store, "e9", "Ricardo", "all is mine");
-    assertAbsent(store, context, "e5, e7", "e8");
-    assertThat(store.size()).isEqualTo(18);
-  }
-
   /**
    * Test that the viewer only takes a snapshot of the passed metadata buffer such that mutations on
    * it do not affect the viewer.
@@ -854,6 +679,7 @@ class StoreTest {
     }
   }
 
+  @Disabled("Till spec change")
   @StoreParameterizedTest
   void writeWithIllegalPosition(Store store) throws IOException {
     try (var editor = edit(store, "e1")) {
@@ -878,6 +704,7 @@ class StoreTest {
     }
   }
 
+  @Disabled("Till spec change")
   @StoreParameterizedTest
   void editorDiscardsWritesAfterClosure(Store store, StoreContext context) throws IOException {
     var editor = edit(store, "e1");
